Snap: Render zoom circle in the center of the camera preview
Previously, the zoom circle was rendered in the center of the entire screen,
making it look unevenly placed due to the uneven sizes of the top and bottom
UI panes. The camera preview in this case is calculated to be the portion
of the screen between the top and bottom UI panes.

Change-Id: I67678f7bc227cf2b7a6eed3161805b2b107c488a