SnapdragonCamera: Allow using an absolute sharpness value
The sharpness value which can be configured normally
is not accurate for most camera HALs.
So, to allow fine tuning these, we introduce
"config_use_absolute_sharpness" which is disabled by
default.
If set to true the sharpness value in "pref_camera_sharpness_default"
is regarded as an absolute sharpness value and
gets passed to the HAL without the addtional
multiplications.

@markakash: Enable it by default

Change-Id: Idff4d5be11b8921e248fece25e1afcb19e9ffb0d
Signed-off-by: Alex Naidis <alex.naidis@linux.com>